@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-		<TargetFrameworks>net6.0-windows;net48</TargetFrameworks>
+		<TargetFrameworks>net8.0-windows;net48</TargetFrameworks>
     <Configurations>Debug;Release;Debug_NuGet;Release_NuGet</Configurations>
   </PropertyGroup>
 
@@ -35,8 +35,8 @@
   </ItemGroup>
     
   <ItemGroup Condition="$(Configuration.Contains('NuGet'))">
-    <PackageReference Include="IdeaStatica.BimApiLink" VersionOverride="25.0.4.1207" />
-    <PackageReference Include="IdeaStatiCa.Plugin" VersionOverride="25.0.4.1207" />
+    <PackageReference Include="IdeaStatica.BimApiLink"/>
+    <PackageReference Include="IdeaStatiCa.Plugin"/>
   </ItemGroup>
       
   <ItemGroup Condition="'$(Configuration)' == 'Debug' Or '$(Configuration)' == 'Release'">
