@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFrameworks>net6.0-windows;net48</TargetFrameworks>
+		<TargetFrameworks>net8.0-windows;net48</TargetFrameworks>
 		<Configurations>Debug;Release;Debug_NuGet;Release_NuGet</Configurations>
 		<UseWPF>true</UseWPF>
 	</PropertyGroup>
