@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFrameworks>net6.0-windows;net48</TargetFrameworks>
+		<TargetFrameworks>net8.0-windows;net48</TargetFrameworks>
 		<Configurations>Debug;Release;Debug_NuGet;Release_NuGet</Configurations>
 		<!--<Nullable>enable</Nullable>-->
 	</PropertyGroup>
@@ -16,7 +16,7 @@
 	</ItemGroup>
 
 	<ItemGroup Condition="$(Configuration.Contains('NuGet'))">
-		<PackageReference Include="IdeaStatiCa.BimApiLink" VersionOverride="25.0.4.1207" />
+		<PackageReference Include="IdeaStatiCa.BimApiLink"/>
 	</ItemGroup>
 
 	<ItemGroup Condition="'$(Configuration)' == 'Debug' Or '$(Configuration)' == 'Release'">
