@@ -1,16 +1,16 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
-		<Nullable>enable</Nullable>
+		<Nullable>disable</Nullable>
 	</PropertyGroup>
 
 	<ItemGroup>
 		
-		<PackageReference Include="IdeaStatiCa.Plugin" VersionOverride="25.0.4.1207" />
-		<PackageReference Include="IdeaStatiCa.PluginLogger" VersionOverride="25.0.4.1207" />
-		<PackageReference Include="IdeaStatiCa.SAF2IOM" VersionOverride="25.0.4.1207" />
+		<PackageReference Include="IdeaStatiCa.Plugin"/>
+		<PackageReference Include="IdeaStatiCa.PluginLogger"/>
+		<PackageReference Include="IdeaStatiCa.SAF2IOM"/>
 	</ItemGroup>
 	
 	
