@@ -1,4 +1,5 @@
-﻿using System;
+﻿using IdeaRS.OpenModel.Geometry3D;
+using System;
 
 namespace IdeaRS.OpenModel.Model
 {
@@ -36,35 +37,14 @@ namespace IdeaRS.OpenModel.Model
 		/// Rotation of Cross-section of Element1D. Difference from default Line LCS
 		/// </summary>
 		public double RotationRx { get; set; }
+		/// <summary>
+		/// Local eccentricity at the begin of Element1D. Used only for export.
+		/// </summary>
+		public Vector3D EccentricityBegin { get; set; } = new Vector3D();
 
 		/// <summary>
-		/// Local eccentricity X at the begin of Element1D
+		/// Local eccentricity at the end of Element1D. Used only for export.
 		/// </summary>
-		public double EccentricityBeginX { get; set; }
-
-		/// <summary>
-		/// Local eccentricity Y at the begin of Element1D
-		/// </summary>
-		public double EccentricityBeginY { get; set; }
-
-		/// <summary>
-		///  Local eccentricity Z at the begin of Element1D
-		/// </summary>
-		public double EccentricityBeginZ { get; set; }
-
-		/// <summary>
-		///  Local eccentricity X at the end of Element1D
-		/// </summary>
-		public double EccentricityEndX { get; set; }
-
-		/// <summary>
-		///  Local eccentricity Y at the end of Element1D
-		/// </summary>
-		public double EccentricityEndY { get; set; }
-
-		/// <summary>
-		///  Local eccentricity Z at the end of Element1D
-		/// </summary>
-		public double EccentricityEndZ { get; set; }
+		public Vector3D EccentricityEnd { get; set; } = new Vector3D();
 	}
 }
