@@ -19,6 +19,7 @@ using System.Xml.Serialization;
 namespace IdeaRS.OpenModel
 {
 	///Version history
+	/// V3.0 24.9.2025 Replaced Eccentricity properties by a single vector. Added CardinalPoint and EccentricityReference to Element1D
 	/// V2.2 19.6.2025 Added new property 'ReferencedGeometry' to 'LoadOnSurface' object
 	/// V2.1 27.08 2024 Added Pins
 	/// V2.0.6 27.08 2024 Grids connectedParts by reference => delete connectedParttIds
@@ -38,7 +39,7 @@ namespace IdeaRS.OpenModel
 		/// </summary>
 		public OpenModel()
 		{
-			Version = "2.2.0";
+			Version = "3.0.0";
 			OriginSettings = null;
 			Point3D = new List<Point3D>();
 			LineSegment3D = new List<LineSegment3D>();
@@ -51,6 +52,7 @@ namespace IdeaRS.OpenModel
 			MatPrestressSteel = new List<MatPrestressSteel>();
 			MatWelding = new List<MatWelding>();
 			MatBoltGrade = new List<MaterialBoltGrade>();
+			MatHeadedStudGrade = new List<MaterialHeadedStudGrade>();
 			CrossSection = new List<CrossSection.CrossSection>();
 			BoltAssembly = new List<BoltAssembly>();
 			Pin = new List<Pin>();
@@ -177,6 +179,11 @@ namespace IdeaRS.OpenModel
 		/// List of BoltGrades
 		/// </summary>
 		public List<MaterialBoltGrade> MatBoltGrade { get; set; }
+
+		/// <summary>
+		/// List of HeadedStudGrades
+		/// </summary>
+		public List<MaterialHeadedStudGrade> MatHeadedStudGrade { get; set; }
 
 		/// <summary>
 		/// List of CrossSection
@@ -692,6 +699,7 @@ namespace IdeaRS.OpenModel
 					{ typeof(MatPrestressSteel).Name, MatPrestressSteel },
 					{ typeof(MatWelding).Name, MatWelding },
 					{ typeof(MaterialBoltGrade).Name, MatBoltGrade },
+					{ typeof(MaterialHeadedStudGrade).Name, MatHeadedStudGrade },
 					{ typeof(CrossSection.CrossSection).Name, CrossSection },
 					{ typeof(CrossSection.ReinforcedCrossSection).Name, ReinforcedCrossSection },
 					{ typeof(BoltAssembly).Name, BoltAssembly },
