@@ -1,4 +1,5 @@
 ﻿using NUnit.Framework;
+using System.Linq;
 using System.Xml.Linq;
 
 namespace IdeaStatiCa.IOM.VersioningServiceTests
@@ -17,8 +18,52 @@ namespace IdeaStatiCa.IOM.VersioningServiceTests
 			IdeaStatiCa.IOM.VersioningServiceTests.UtHelper.Attach(expectedXml, $"{testCaseName}-expected.xml");
 			IdeaStatiCa.IOM.VersioningServiceTests.UtHelper.Attach(actualXml, $"{testCaseName}-actual.xml");
 
-			Assert.IsTrue(XNode.DeepEquals(XElement.Parse(expectedXml), XElement.Parse(actualXml)),
-				String.Format("{0} \n does not equal \n{1}", actualXml, expectedXml));
+			var expected = XElement.Parse(expectedXml);
+			Normalize(expected);
+
+			var actual = XElement.Parse(actualXml);
+			Normalize(actual);
+
+			Assert.IsTrue(XNode.DeepEquals(expected, actual),
+				string.Format("{0} \n does not equal \n{1}", actualXml, expectedXml));
+		}
+
+		private static void Normalize(XElement element)
+		{
+			// sort attributes
+			List<XAttribute> orderedAttributes = element.Attributes()
+				.OrderBy(x => x.Name.ToString())
+				.ToList();
+			
+			element.ReplaceAttributes(orderedAttributes);
+
+			// recursively sort children
+			List<XNode> orderedChildren = element.Nodes()
+				.OrderBy(x => x, Comparer<XNode>.Create(CompareNodes))
+				.ToList();
+			
+			foreach (var child in orderedChildren.OfType<XElement>())
+			{
+				Normalize(child);
+			}
+
+			element.ReplaceNodes(orderedChildren);
+		}
+
+		private static int CompareNodes(XNode x, XNode y)
+		{
+			if(x is XElement xe && y is XElement ye)
+			{
+				return string.Compare(xe.Name.ToString(), ye.Name.ToString(), StringComparison.Ordinal);
+			}
+			else if (x is XText xt && y is XText yt)
+			{
+				return string.Compare(xt.Value, yt.Value, StringComparison.Ordinal);
+			}
+			else
+			{
+				return 0;
+			}
 		}
 	}
 }
