@@ -22,7 +22,7 @@ namespace IdeaConWpfApp
 		{
 			InitializeComponent();
 			ProjectsList.ItemsSource = projectFiles;
-			IdeaPathText.Text = ideaPath = @"C:\Program Files\IDEA StatiCa\StatiCa 25.0\";
+			IdeaPathText.Text = ideaPath = @"C:\Program Files\IDEA StatiCa\StatiCa 25.1\";
 			DesignCodeComboBox.SelectedIndex = 0;
 			this.Closed += MainWindow_Closed;
 		}
@@ -38,7 +38,7 @@ namespace IdeaConWpfApp
 			var dialog = new System.Windows.Forms.FolderBrowserDialog()
 			{
 				Description = "Select folder where is API v25",
-				SelectedPath = @"C:\Program Files\IDEA StatiCa\StatiCa 25.0\"
+				SelectedPath = @"C:\Program Files\IDEA StatiCa\StatiCa 25.1\"
 			};
 
 			if (dialog.ShowDialog() == System.Windows.Forms.DialogResult.OK)
@@ -53,7 +53,7 @@ namespace IdeaConWpfApp
 			var dialog = new System.Windows.Forms.FolderBrowserDialog()
 			{
 				Description = "Select folder with .ideaCon files",
-				SelectedPath = selectedFolderPath // Default to last selected folder
+				SelectedPath = selectedFolderPath! // Default to last selected folder
 			};
 
 			if (dialog.ShowDialog() == System.Windows.Forms.DialogResult.OK)
