@@ -34,28 +34,28 @@ namespace IdeaStatiCa.ConnectionApi.Api
     {
         #region Synchronous Operations
         /// <summary>
-        /// Run CBFEM caluclation and return the summary of the results
+        /// 
         /// </summary>
         /// <exception cref="IdeaStatiCa.ConnectionApi.Client.ApiException">Thrown when fails to make API call</exception>
-        /// <param name="projectId">The unique identifier of the opened project in the ConnectionRestApi service</param>
-        /// <param name="conCalculationParameter">List of connections to calculate and a type of CBFEM analysis</param>
+        /// <param name="projectId"></param>
+        /// <param name="requestBody"></param>
         /// <param name="operationIndex">Index associated with the operation.</param>
         /// <returns>List&lt;ConResultSummary&gt;</returns>
-        List<ConResultSummary> Calculate(Guid projectId, ConCalculationParameter conCalculationParameter, int operationIndex = 0);
+        List<ConResultSummary> Calculate(Guid projectId, List<int> requestBody, int operationIndex = 0);
 
         /// <summary>
-        /// Run CBFEM caluclation and return the summary of the results
+        /// 
         /// </summary>
         /// <remarks>
         /// 
         /// </remarks>
         /// <exception cref="IdeaStatiCa.ConnectionApi.Client.ApiException">Thrown when fails to make API call</exception>
-        /// <param name="projectId">The unique identifier of the opened project in the ConnectionRestApi service</param>
-/// <param name="conCalculationParameter">List of connections to calculate and a type of CBFEM analysis</param>
+        /// <param name="projectId"></param>
+/// <param name="requestBody"></param>
         /// <param name="requestedType">Requested content type in the response.</param>        
         /// <param name="operationIndex">Index associated with the operation.</param>
         /// <returns>ApiResponse of List&lt;ConResultSummary&gt;</returns>
-        ApiResponse<List<ConResultSummary>> CalculateWithHttpInfo(Guid projectId, ConCalculationParameter conCalculationParameter, string requestedType = null, int operationIndex = 0);
+        ApiResponse<List<ConResultSummary>> CalculateWithHttpInfo(Guid projectId, List<int> requestBody, string requestedType = null, int operationIndex = 0);
         /// <summary>
         /// Get json string which represents raw CBFEM results (an instance of CheckResultsData)
         /// </summary>
@@ -84,10 +84,10 @@ namespace IdeaStatiCa.ConnectionApi.Api
         /// </summary>
         /// <exception cref="IdeaStatiCa.ConnectionApi.Client.ApiException">Thrown when fails to make API call</exception>
         /// <param name="projectId">The unique identifier of the opened project in the ConnectionRestApi service</param>
-        /// <param name="conCalculationParameter">List of connections to calculate and a type of CBFEM analysis (optional)</param>
+        /// <param name="requestBody">List of connections to calculate and a type of CBFEM analysis</param>
         /// <param name="operationIndex">Index associated with the operation.</param>
         /// <returns>List&lt;ConnectionCheckRes&gt;</returns>
-        List<ConnectionCheckRes> GetResults(Guid projectId, ConCalculationParameter conCalculationParameter = default(ConCalculationParameter), int operationIndex = 0);
+        List<ConnectionCheckRes> GetResults(Guid projectId, List<int> requestBody, int operationIndex = 0);
 
         /// <summary>
         /// Get detailed results of the CBFEM analysis
@@ -97,11 +97,11 @@ namespace IdeaStatiCa.ConnectionApi.Api
         /// </remarks>
         /// <exception cref="IdeaStatiCa.ConnectionApi.Client.ApiException">Thrown when fails to make API call</exception>
         /// <param name="projectId">The unique identifier of the opened project in the ConnectionRestApi service</param>
-/// <param name="conCalculationParameter">List of connections to calculate and a type of CBFEM analysis (optional)</param>
+/// <param name="requestBody">List of connections to calculate and a type of CBFEM analysis</param>
         /// <param name="requestedType">Requested content type in the response.</param>        
         /// <param name="operationIndex">Index associated with the operation.</param>
         /// <returns>ApiResponse of List&lt;ConnectionCheckRes&gt;</returns>
-        ApiResponse<List<ConnectionCheckRes>> GetResultsWithHttpInfo(Guid projectId, ConCalculationParameter conCalculationParameter = default(ConCalculationParameter), string requestedType = null, int operationIndex = 0);
+        ApiResponse<List<ConnectionCheckRes>> GetResultsWithHttpInfo(Guid projectId, List<int> requestBody, string requestedType = null, int operationIndex = 0);
         #endregion Synchronous Operations
     }
 
@@ -112,33 +112,33 @@ namespace IdeaStatiCa.ConnectionApi.Api
     {
         #region Asynchronous Operations
         /// <summary>
-        /// Run CBFEM caluclation and return the summary of the results
+        /// 
         /// </summary>
         /// <remarks>
         /// 
         /// </remarks>
         /// <exception cref="IdeaStatiCa.ConnectionApi.Client.ApiException">Thrown when fails to make API call</exception>
-        /// <param name="projectId">The unique identifier of the opened project in the ConnectionRestApi service</param>
-        /// <param name="conCalculationParameter">List of connections to calculate and a type of CBFEM analysis</param>
+        /// <param name="projectId"></param>
+        /// <param name="requestBody"></param>
         /// <param name="operationIndex">Index associated with the operation.</param>
         /// <param name="cancellationToken">Cancellation Token to cancel the request.</param>
         /// <returns>Task of List&lt;ConResultSummary&gt;</returns>
-        System.Threading.Tasks.Task<List<ConResultSummary>> CalculateAsync(Guid projectId, ConCalculationParameter conCalculationParameter, int operationIndex = 0, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken));
+        System.Threading.Tasks.Task<List<ConResultSummary>> CalculateAsync(Guid projectId, List<int> requestBody, int operationIndex = 0, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken));
 
         /// <summary>
-        /// Run CBFEM caluclation and return the summary of the results
+        /// 
         /// </summary>
         /// <remarks>
         /// 
         /// </remarks>
         /// <exception cref="IdeaStatiCa.ConnectionApi.Client.ApiException">Thrown when fails to make API call</exception>
-        /// <param name="projectId">The unique identifier of the opened project in the ConnectionRestApi service</param>
-        /// <param name="conCalculationParameter">List of connections to calculate and a type of CBFEM analysis</param>
+        /// <param name="projectId"></param>
+        /// <param name="requestBody"></param>
         /// <param name="requestedType">Requested content type in the response.</param>
         /// <param name="operationIndex">Index associated with the operation.</param>
         /// <param name="cancellationToken">Cancellation Token to cancel the request.</param>
         /// <returns>Task of ApiResponse (List&lt;ConResultSummary&gt;)</returns>
-        System.Threading.Tasks.Task<ApiResponse<List<ConResultSummary>>> CalculateWithHttpInfoAsync(Guid projectId, ConCalculationParameter conCalculationParameter, string requestedType = null, int operationIndex = 0, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken));
+        System.Threading.Tasks.Task<ApiResponse<List<ConResultSummary>>> CalculateWithHttpInfoAsync(Guid projectId, List<int> requestBody, string requestedType = null, int operationIndex = 0, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken));
         /// <summary>
         /// Get json string which represents raw CBFEM results (an instance of CheckResultsData)
         /// </summary>
@@ -175,11 +175,11 @@ namespace IdeaStatiCa.ConnectionApi.Api
         /// </remarks>
         /// <exception cref="IdeaStatiCa.ConnectionApi.Client.ApiException">Thrown when fails to make API call</exception>
         /// <param name="projectId">The unique identifier of the opened project in the ConnectionRestApi service</param>
-        /// <param name="conCalculationParameter">List of connections to calculate and a type of CBFEM analysis (optional)</param>
+        /// <param name="requestBody">List of connections to calculate and a type of CBFEM analysis</param>
         /// <param name="operationIndex">Index associated with the operation.</param>
         /// <param name="cancellationToken">Cancellation Token to cancel the request.</param>
         /// <returns>Task of List&lt;ConnectionCheckRes&gt;</returns>
-        System.Threading.Tasks.Task<List<ConnectionCheckRes>> GetResultsAsync(Guid projectId, ConCalculationParameter conCalculationParameter = default(ConCalculationParameter), int operationIndex = 0, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken));
+        System.Threading.Tasks.Task<List<ConnectionCheckRes>> GetResultsAsync(Guid projectId, List<int> requestBody, int operationIndex = 0, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken));
 
         /// <summary>
         /// Get detailed results of the CBFEM analysis
@@ -189,12 +189,12 @@ namespace IdeaStatiCa.ConnectionApi.Api
         /// </remarks>
         /// <exception cref="IdeaStatiCa.ConnectionApi.Client.ApiException">Thrown when fails to make API call</exception>
         /// <param name="projectId">The unique identifier of the opened project in the ConnectionRestApi service</param>
-        /// <param name="conCalculationParameter">List of connections to calculate and a type of CBFEM analysis (optional)</param>
+        /// <param name="requestBody">List of connections to calculate and a type of CBFEM analysis</param>
         /// <param name="requestedType">Requested content type in the response.</param>
         /// <param name="operationIndex">Index associated with the operation.</param>
         /// <param name="cancellationToken">Cancellation Token to cancel the request.</param>
         /// <returns>Task of ApiResponse (List&lt;ConnectionCheckRes&gt;)</returns>
-        System.Threading.Tasks.Task<ApiResponse<List<ConnectionCheckRes>>> GetResultsWithHttpInfoAsync(Guid projectId, ConCalculationParameter conCalculationParameter = default(ConCalculationParameter), string requestedType = null, int operationIndex = 0, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken));
+        System.Threading.Tasks.Task<ApiResponse<List<ConnectionCheckRes>>> GetResultsWithHttpInfoAsync(Guid projectId, List<int> requestBody, string requestedType = null, int operationIndex = 0, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken));
         #endregion Asynchronous Operations
     }
 
@@ -316,34 +316,34 @@ namespace IdeaStatiCa.ConnectionApi.Api
         }
 
         /// <summary>
-        /// Run CBFEM caluclation and return the summary of the results 
+        ///  
         /// </summary>
         /// <exception cref="IdeaStatiCa.ConnectionApi.Client.ApiException">Thrown when fails to make API call</exception>
-        /// <param name="projectId">The unique identifier of the opened project in the ConnectionRestApi service</param>
-        /// <param name="conCalculationParameter">List of connections to calculate and a type of CBFEM analysis</param>
+        /// <param name="projectId"></param>
+        /// <param name="requestBody"></param>
         /// <param name="operationIndex">Index associated with the operation.</param>
         /// <returns>List&lt;ConResultSummary&gt;</returns>
-        public List<ConResultSummary> Calculate(Guid projectId, ConCalculationParameter conCalculationParameter, int operationIndex = 0)
+        public List<ConResultSummary> Calculate(Guid projectId, List<int> requestBody, int operationIndex = 0)
         {
-            IdeaStatiCa.ConnectionApi.Client.ApiResponse<List<ConResultSummary>> localVarResponse = CalculateWithHttpInfo(projectId, conCalculationParameter);
+            IdeaStatiCa.ConnectionApi.Client.ApiResponse<List<ConResultSummary>> localVarResponse = CalculateWithHttpInfo(projectId, requestBody);
             return localVarResponse.Data;
         }
 
         /// <summary>
-        /// Run CBFEM caluclation and return the summary of the results 
+        ///  
         /// </summary>
         /// <exception cref="IdeaStatiCa.ConnectionApi.Client.ApiException">Thrown when fails to make API call</exception>
-        /// <param name="projectId">The unique identifier of the opened project in the ConnectionRestApi service</param>
-/// <param name="conCalculationParameter">List of connections to calculate and a type of CBFEM analysis</param>
+        /// <param name="projectId"></param>
+/// <param name="requestBody"></param>
         /// <param name="requestedType">Requested content type in the response.</param>
         /// <param name="operationIndex">Index associated with the operation.</param>
         /// <returns>ApiResponse of List&lt;ConResultSummary&gt;</returns>
-        public IdeaStatiCa.ConnectionApi.Client.ApiResponse<List<ConResultSummary>> CalculateWithHttpInfo(Guid projectId, ConCalculationParameter conCalculationParameter, string requestedType = null, int operationIndex = 0)
+        public IdeaStatiCa.ConnectionApi.Client.ApiResponse<List<ConResultSummary>> CalculateWithHttpInfo(Guid projectId, List<int> requestBody, string requestedType = null, int operationIndex = 0)
         {
-            // verify the required parameter 'conCalculationParameter' is set
-            if (conCalculationParameter == null)
+            // verify the required parameter 'requestBody' is set
+            if (requestBody == null)
             {
-                throw new IdeaStatiCa.ConnectionApi.Client.ApiException(400, "Missing required parameter 'conCalculationParameter' when calling CalculationApi->Calculate");
+                throw new IdeaStatiCa.ConnectionApi.Client.ApiException(400, "Missing required parameter 'requestBody' when calling CalculationApi->Calculate");
             }
 
             IdeaStatiCa.ConnectionApi.Client.RequestOptions localVarRequestOptions = new IdeaStatiCa.ConnectionApi.Client.RequestOptions();
@@ -375,7 +375,7 @@ namespace IdeaStatiCa.ConnectionApi.Api
             }
 
             localVarRequestOptions.PathParameters.Add("projectId", IdeaStatiCa.ConnectionApi.Client.ClientUtils.ParameterToString(projectId)); // path parameter
-            localVarRequestOptions.Data = conCalculationParameter;
+            localVarRequestOptions.Data = requestBody;
 
             localVarRequestOptions.Operation = "CalculationApi.Calculate";
             localVarRequestOptions.OperationIndex = operationIndex;
@@ -396,36 +396,36 @@ namespace IdeaStatiCa.ConnectionApi.Api
         }
 
         /// <summary>
-        /// Run CBFEM caluclation and return the summary of the results 
+        ///  
         /// </summary>
         /// <exception cref="IdeaStatiCa.ConnectionApi.Client.ApiException">Thrown when fails to make API call</exception>
-        /// <param name="projectId">The unique identifier of the opened project in the ConnectionRestApi service</param>
-        /// <param name="conCalculationParameter">List of connections to calculate and a type of CBFEM analysis</param>
+        /// <param name="projectId"></param>
+        /// <param name="requestBody"></param>
         /// <param name="operationIndex">Index associated with the operation.</param>
         /// <param name="cancellationToken">Cancellation Token to cancel the request.</param>
         /// <returns>Task of List&lt;ConResultSummary&gt;</returns>
-        public async System.Threading.Tasks.Task<List<ConResultSummary>> CalculateAsync(Guid projectId, ConCalculationParameter conCalculationParameter, int operationIndex = 0, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken))
+        public async System.Threading.Tasks.Task<List<ConResultSummary>> CalculateAsync(Guid projectId, List<int> requestBody, int operationIndex = 0, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken))
         {
-            IdeaStatiCa.ConnectionApi.Client.ApiResponse<List<ConResultSummary>> localVarResponse = await CalculateWithHttpInfoAsync(projectId, conCalculationParameter, null, operationIndex, cancellationToken).ConfigureAwait(false);
+            IdeaStatiCa.ConnectionApi.Client.ApiResponse<List<ConResultSummary>> localVarResponse = await CalculateWithHttpInfoAsync(projectId, requestBody, null, operationIndex, cancellationToken).ConfigureAwait(false);
             return localVarResponse.Data;
         }
 
         /// <summary>
-        /// Run CBFEM caluclation and return the summary of the results 
+        ///  
         /// </summary>
         /// <exception cref="IdeaStatiCa.ConnectionApi.Client.ApiException">Thrown when fails to make API call</exception>
-        /// <param name="projectId">The unique identifier of the opened project in the ConnectionRestApi service</param>
-        /// <param name="conCalculationParameter">List of connections to calculate and a type of CBFEM analysis</param>
+        /// <param name="projectId"></param>
+        /// <param name="requestBody"></param>
         /// <param name="requestedType">Requested content type in the response.</param>
         /// <param name="operationIndex">Index associated with the operation.</param>
         /// <param name="cancellationToken">Cancellation Token to cancel the request.</param>
         /// <returns>Task of ApiResponse (List&lt;ConResultSummary&gt;)</returns>
-        public async System.Threading.Tasks.Task<IdeaStatiCa.ConnectionApi.Client.ApiResponse<List<ConResultSummary>>> CalculateWithHttpInfoAsync(Guid projectId, ConCalculationParameter conCalculationParameter, string requestedType = null, int operationIndex = 0, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken))
+        public async System.Threading.Tasks.Task<IdeaStatiCa.ConnectionApi.Client.ApiResponse<List<ConResultSummary>>> CalculateWithHttpInfoAsync(Guid projectId, List<int> requestBody, string requestedType = null, int operationIndex = 0, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken))
         {
-            // verify the required parameter 'conCalculationParameter' is set
-            if (conCalculationParameter == null)
+            // verify the required parameter 'requestBody' is set
+            if (requestBody == null)
             {
-                throw new IdeaStatiCa.ConnectionApi.Client.ApiException(400, "Missing required parameter 'conCalculationParameter' when calling CalculationApi->Calculate");
+                throw new IdeaStatiCa.ConnectionApi.Client.ApiException(400, "Missing required parameter 'requestBody' when calling CalculationApi->Calculate");
             }
 
 
@@ -458,7 +458,7 @@ namespace IdeaStatiCa.ConnectionApi.Api
             }
 
             localVarRequestOptions.PathParameters.Add("projectId", IdeaStatiCa.ConnectionApi.Client.ClientUtils.ParameterToString(projectId)); // path parameter
-            localVarRequestOptions.Data = conCalculationParameter;
+            localVarRequestOptions.Data = requestBody;
 
             localVarRequestOptions.Operation = "CalculationApi.Calculate";
             localVarRequestOptions.OperationIndex = operationIndex;
@@ -636,12 +636,12 @@ namespace IdeaStatiCa.ConnectionApi.Api
         /// </summary>
         /// <exception cref="IdeaStatiCa.ConnectionApi.Client.ApiException">Thrown when fails to make API call</exception>
         /// <param name="projectId">The unique identifier of the opened project in the ConnectionRestApi service</param>
-        /// <param name="conCalculationParameter">List of connections to calculate and a type of CBFEM analysis (optional)</param>
+        /// <param name="requestBody">List of connections to calculate and a type of CBFEM analysis</param>
         /// <param name="operationIndex">Index associated with the operation.</param>
         /// <returns>List&lt;ConnectionCheckRes&gt;</returns>
-        public List<ConnectionCheckRes> GetResults(Guid projectId, ConCalculationParameter conCalculationParameter = default(ConCalculationParameter), int operationIndex = 0)
+        public List<ConnectionCheckRes> GetResults(Guid projectId, List<int> requestBody, int operationIndex = 0)
         {
-            IdeaStatiCa.ConnectionApi.Client.ApiResponse<List<ConnectionCheckRes>> localVarResponse = GetResultsWithHttpInfo(projectId, conCalculationParameter);
+            IdeaStatiCa.ConnectionApi.Client.ApiResponse<List<ConnectionCheckRes>> localVarResponse = GetResultsWithHttpInfo(projectId, requestBody);
             return localVarResponse.Data;
         }
 
@@ -650,12 +650,18 @@ namespace IdeaStatiCa.ConnectionApi.Api
         /// </summary>
         /// <exception cref="IdeaStatiCa.ConnectionApi.Client.ApiException">Thrown when fails to make API call</exception>
         /// <param name="projectId">The unique identifier of the opened project in the ConnectionRestApi service</param>
-/// <param name="conCalculationParameter">List of connections to calculate and a type of CBFEM analysis (optional)</param>
+/// <param name="requestBody">List of connections to calculate and a type of CBFEM analysis</param>
         /// <param name="requestedType">Requested content type in the response.</param>
         /// <param name="operationIndex">Index associated with the operation.</param>
         /// <returns>ApiResponse of List&lt;ConnectionCheckRes&gt;</returns>
-        public IdeaStatiCa.ConnectionApi.Client.ApiResponse<List<ConnectionCheckRes>> GetResultsWithHttpInfo(Guid projectId, ConCalculationParameter conCalculationParameter = default(ConCalculationParameter), string requestedType = null, int operationIndex = 0)
+        public IdeaStatiCa.ConnectionApi.Client.ApiResponse<List<ConnectionCheckRes>> GetResultsWithHttpInfo(Guid projectId, List<int> requestBody, string requestedType = null, int operationIndex = 0)
         {
+            // verify the required parameter 'requestBody' is set
+            if (requestBody == null)
+            {
+                throw new IdeaStatiCa.ConnectionApi.Client.ApiException(400, "Missing required parameter 'requestBody' when calling CalculationApi->GetResults");
+            }
+
             IdeaStatiCa.ConnectionApi.Client.RequestOptions localVarRequestOptions = new IdeaStatiCa.ConnectionApi.Client.RequestOptions();
 
             string[] _contentTypes = new string[] {
@@ -685,7 +691,7 @@ namespace IdeaStatiCa.ConnectionApi.Api
             }
 
             localVarRequestOptions.PathParameters.Add("projectId", IdeaStatiCa.ConnectionApi.Client.ClientUtils.ParameterToString(projectId)); // path parameter
-            localVarRequestOptions.Data = conCalculationParameter;
+            localVarRequestOptions.Data = requestBody;
 
             localVarRequestOptions.Operation = "CalculationApi.GetResults";
             localVarRequestOptions.OperationIndex = operationIndex;
@@ -710,13 +716,13 @@ namespace IdeaStatiCa.ConnectionApi.Api
         /// </summary>
         /// <exception cref="IdeaStatiCa.ConnectionApi.Client.ApiException">Thrown when fails to make API call</exception>
         /// <param name="projectId">The unique identifier of the opened project in the ConnectionRestApi service</param>
-        /// <param name="conCalculationParameter">List of connections to calculate and a type of CBFEM analysis (optional)</param>
+        /// <param name="requestBody">List of connections to calculate and a type of CBFEM analysis</param>
         /// <param name="operationIndex">Index associated with the operation.</param>
         /// <param name="cancellationToken">Cancellation Token to cancel the request.</param>
         /// <returns>Task of List&lt;ConnectionCheckRes&gt;</returns>
-        public async System.Threading.Tasks.Task<List<ConnectionCheckRes>> GetResultsAsync(Guid projectId, ConCalculationParameter conCalculationParameter = default(ConCalculationParameter), int operationIndex = 0, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken))
+        public async System.Threading.Tasks.Task<List<ConnectionCheckRes>> GetResultsAsync(Guid projectId, List<int> requestBody, int operationIndex = 0, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken))
         {
-            IdeaStatiCa.ConnectionApi.Client.ApiResponse<List<ConnectionCheckRes>> localVarResponse = await GetResultsWithHttpInfoAsync(projectId, conCalculationParameter, null, operationIndex, cancellationToken).ConfigureAwait(false);
+            IdeaStatiCa.ConnectionApi.Client.ApiResponse<List<ConnectionCheckRes>> localVarResponse = await GetResultsWithHttpInfoAsync(projectId, requestBody, null, operationIndex, cancellationToken).ConfigureAwait(false);
             return localVarResponse.Data;
         }
 
@@ -725,13 +731,19 @@ namespace IdeaStatiCa.ConnectionApi.Api
         /// </summary>
         /// <exception cref="IdeaStatiCa.ConnectionApi.Client.ApiException">Thrown when fails to make API call</exception>
         /// <param name="projectId">The unique identifier of the opened project in the ConnectionRestApi service</param>
-        /// <param name="conCalculationParameter">List of connections to calculate and a type of CBFEM analysis (optional)</param>
+        /// <param name="requestBody">List of connections to calculate and a type of CBFEM analysis</param>
         /// <param name="requestedType">Requested content type in the response.</param>
         /// <param name="operationIndex">Index associated with the operation.</param>
         /// <param name="cancellationToken">Cancellation Token to cancel the request.</param>
         /// <returns>Task of ApiResponse (List&lt;ConnectionCheckRes&gt;)</returns>
-        public async System.Threading.Tasks.Task<IdeaStatiCa.ConnectionApi.Client.ApiResponse<List<ConnectionCheckRes>>> GetResultsWithHttpInfoAsync(Guid projectId, ConCalculationParameter conCalculationParameter = default(ConCalculationParameter), string requestedType = null, int operationIndex = 0, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken))
+        public async System.Threading.Tasks.Task<IdeaStatiCa.ConnectionApi.Client.ApiResponse<List<ConnectionCheckRes>>> GetResultsWithHttpInfoAsync(Guid projectId, List<int> requestBody, string requestedType = null, int operationIndex = 0, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken))
         {
+            // verify the required parameter 'requestBody' is set
+            if (requestBody == null)
+            {
+                throw new IdeaStatiCa.ConnectionApi.Client.ApiException(400, "Missing required parameter 'requestBody' when calling CalculationApi->GetResults");
+            }
+
 
             IdeaStatiCa.ConnectionApi.Client.RequestOptions localVarRequestOptions = new IdeaStatiCa.ConnectionApi.Client.RequestOptions();
 
@@ -762,7 +774,7 @@ namespace IdeaStatiCa.ConnectionApi.Api
             }
 
             localVarRequestOptions.PathParameters.Add("projectId", IdeaStatiCa.ConnectionApi.Client.ClientUtils.ParameterToString(projectId)); // path parameter
-            localVarRequestOptions.Data = conCalculationParameter;
+            localVarRequestOptions.Data = requestBody;
 
             localVarRequestOptions.Operation = "CalculationApi.GetResults";
             localVarRequestOptions.OperationIndex = operationIndex;
