@@ -44,11 +44,12 @@ with connection_api_service_attacher.ConnectionApiServiceAttacher(baseUrl).creat
 
         # run stress-strain CBFEM analysis for the connection id = 1
         calcParams = ideastatica_connection_api.ConCalculationParameter() # ConCalculationParameter | List of connections to calculate and a type of CBFEM analysis (optional)
-        calcParams.connection_ids = [connection1.id]
-        calcParams.analysis_type = "stress_Strain"
+        calcParams.connection_ids = [connection1.id]    # calculate only connection1
+        calcParams.analysis_type = "stress_Strain"      # obsolete solution (the analysis_type is taked from the connection now)
   
+
         # run stress-strain analysis for connection1
-        con1_cbfem_results = api_client.calculation.calculate(api_client.project.active_project_id, calcParams)
+        con1_cbfem_results = api_client.calculation.calculate(api_client.project.active_project_id, calcParams.connection_ids)
 
         #results stress-strain analysis
         pprint(con1_cbfem_results)
@@ -57,17 +58,27 @@ with connection_api_service_attacher.ConnectionApiServiceAttacher(baseUrl).creat
         pprint(detailed_results)
 
         # run Fatigue analysis for connection1
-        calcParams.connection_ids = [connection1.id]
-        calcParams.analysis_type = "fatigues"    
 
-        # run stress-strain analysis for the connection
-        con1_cbfem_results = api_client.calculation.calculate(api_client.project.active_project_id, calcParams)
+        # set analysis type stress-strain
+        calcParams.analysis_type = "fatigues"
+        connection1.analysis_type = calcParams.analysis_type
+        updated_connection1 = api_client.connection.update_connection(api_client.project.active_project_id, connection1.id, connection1)
+
+        if(connection1.analysis_type != updated_connection1.analysis_type):
+            raise ValueError("Connection analysis type was not updated successfully.")
+
+        # re-run analysis for the connection - now analysis_type is "fatigues"
+        con1_cbfem_results = api_client.calculation.calculate(api_client.project.active_project_id, calcParams.connection_ids)
 
         # results Fatigue analysis
         pprint(con1_cbfem_results)
 
-        detailed_results = api_client.calculation.get_results(api_client.project.active_project_id, calcParams)
-        pprint(detailed_results)
+        fatiguedetailed_results = api_client.calculation.get_results(api_client.project.active_project_id, calcParams)
+        pprint(fatiguedetailed_results)
+
+        results_text = api_client.calculation.get_raw_json_results(api_client.project.active_project_id, calcParams)
+        firstConnectionRawResult = results_text[0]
+        pprint(firstConnectionRawResult)
 
     except Exception as e:
         print("Operation failed : %s\n" % e)
