@@ -35,6 +35,7 @@ from ideastatica_connection_api.models.con_analysis_type_enum import ConAnalysis
 from ideastatica_connection_api.models.con_calculation_parameter import ConCalculationParameter
 from ideastatica_connection_api.models.con_connection import ConConnection
 from ideastatica_connection_api.models.con_conversion_settings import ConConversionSettings
+from ideastatica_connection_api.models.con_design_set_type import ConDesignSetType
 from ideastatica_connection_api.models.con_load_effect import ConLoadEffect
 from ideastatica_connection_api.models.con_load_effect_member_load import ConLoadEffectMemberLoad
 from ideastatica_connection_api.models.con_load_effect_position_enum import ConLoadEffectPositionEnum
@@ -57,16 +58,14 @@ from ideastatica_connection_api.models.con_project_data import ConProjectData
 from ideastatica_connection_api.models.con_result_summary import ConResultSummary
 from ideastatica_connection_api.models.con_template_apply_param import ConTemplateApplyParam
 from ideastatica_connection_api.models.con_template_mapping_get_param import ConTemplateMappingGetParam
+from ideastatica_connection_api.models.con_template_publish_param import ConTemplatePublishParam
+from ideastatica_connection_api.models.con_weld_sizing_method_enum import ConWeldSizingMethodEnum
 from ideastatica_connection_api.models.concrete_block import ConcreteBlock
 from ideastatica_connection_api.models.concrete_block_data import ConcreteBlockData
-from ideastatica_connection_api.models.concrete_setup import ConcreteSetup
-from ideastatica_connection_api.models.cone_breakout_check_type import ConeBreakoutCheckType
 from ideastatica_connection_api.models.connection_check_res import ConnectionCheckRes
 from ideastatica_connection_api.models.connection_data import ConnectionData
-from ideastatica_connection_api.models.connection_setup import ConnectionSetup
 from ideastatica_connection_api.models.conversion_mapping import ConversionMapping
 from ideastatica_connection_api.models.country_code import CountryCode
-from ideastatica_connection_api.models.crt_comp_check_is import CrtCompCheckIS
 from ideastatica_connection_api.models.cut_beam_by_beam_data import CutBeamByBeamData
 from ideastatica_connection_api.models.cut_data import CutData
 from ideastatica_connection_api.models.cut_method import CutMethod
@@ -101,5 +100,4 @@ from ideastatica_connection_api.models.text import Text
 from ideastatica_connection_api.models.text_position import TextPosition
 from ideastatica_connection_api.models.vector3_d import Vector3D
 from ideastatica_connection_api.models.weld_data import WeldData
-from ideastatica_connection_api.models.weld_evaluation import WeldEvaluation
 from ideastatica_connection_api.models.weld_type import WeldType
