@@ -109,7 +109,7 @@ class ApiClient:
             self.default_headers[header_name] = header_value
         self.cookie = cookie
         # Set default User-Agent.
-        self.user_agent = 'OpenAPI-Generator/25.0.1.1008/python'
+        self.user_agent = 'OpenAPI-Generator/25.1.0.2640/python'
         self.client_side_validation = configuration.client_side_validation
 
     def __enter__(self):
