@@ -1,4 +1,5 @@
-﻿using IdeaStatiCa.BimApi;
+﻿using IdeaRS.OpenModel.Model;
+using IdeaStatiCa.BimApi;
 using IdeaStatiCa.BimApi.Results;
 using System.Collections.Generic;
 using System.Linq;
@@ -16,10 +17,6 @@ namespace IdeaStatiCa.RamToIdea.BimApi
 		public IIdeaCrossSection StartCrossSection { get; set; }
 
 		public IIdeaCrossSection EndCrossSection { get; set; }
-
-		public IdeaVector3D EccentricityBegin { get; set; } = new IdeaVector3D(0, 0, 0);
-
-		public IdeaVector3D EccentricityEnd { get; set; } = new IdeaVector3D(0, 0, 0);
 
 		public double RotationRx { get; set; }
 
