@@ -12,12 +12,12 @@ namespace IdeaStatiCa.RamToIdea.Sections
 
 		public CrossSectionType Type => _cssParams.CrossSectionType;
 
-		public HashSet<Parameter> Parameters => new HashSet<Parameter>(_cssParams.Parameters);
+		public HashSet<Parameter> Parameters => new(_cssParams.Parameters);
 
 		private readonly CrossSectionParameter _cssParams;
 
-		public RamSectionParametric(IMaterialFactory materialFactory, double height, RamMemberProperties props, CrossSectionParameter cssParam)
-			: base(materialFactory, height, props)
+		public RamSectionParametric(IMaterialFactory materialFactory, RamMemberProperties props, CrossSectionParameter cssParam)
+			: base(materialFactory, props)
 		{
 			_cssParams = cssParam;
 		}
