@@ -151,7 +151,7 @@ namespace IdeaStatiCa.TeklaStructuresPlugin
 				}
 				if (!File.Exists(checkbotLocation))
 				{
-					//checkbot app is in net48 folder and teklaPlugin is in net6.0-windows  of setup
+					//checkbot app is in net48 folder and teklaPlugin is in net8.0-windows  of setup
 					checkbotLocation = Path.Combine(checkbotRoot, "..\\net48", IdeaStatiCa.Plugin.Constants.CheckbotAppName);
 				}
 
